+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [NEXT] -
+
+### Add
+
+- `set` method on `collSpecable`;
+
+---
+
 ## [0.0.1] - 2021-03-30
 
 Initial working code.